pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
name = "acbm"
version = "0.1.0"
authors = ["Hussein Mahfouz <husseinmahfouz93@gmail.com>"]
homepage = "https://github.com/alan-turing-institute/acbm"
repository = "https://github.com/alan-turing-institute/acbm"
license = "Apache-2.0"
description = "A package to create activity-based models (for transport demand modelling)"
readme = "README.md"

classifiers = [
  "Development Status :: 1 - Planning",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering",

]

[tool.poetry.dependencies]
python = "^3.10"

pytest = { version = ">=6", optional = true }
pytest-cov = { version = ">=3", optional = true }
pandas = "^2.2.0"
uatk-spc = { git = "https://github.com/alan-turing-institute/uatk-spc.git", subdirectory = "python" }
geopandas = "^0.14.3"
matplotlib = "^3.8.3"
scikit-learn = "^1.4.1.post1"
seaborn = "^0.13.2"
pandarallel = "^1.6.5"
ipywidgets = "^8.1.2"
tqdm = "^4.66.2"
pulp = "^2.8.0"
libpysal = "^4.11.0"
click = "^8.1.7"
tomlkit = "^0.13.0"
cml-pam = "0.3.2"
gdal = "<=3.8.4"
pandera = "^0.20.4"
osmox = { git = "https://github.com/arup-group/osmox" }
pyrosm = "^0.6.2"
jsonschema = "^4.23.0"
jcs = "^0.2.1"

[tool.poetry.dev-dependencies]
pytest = ">= 6"
pytest-cov = ">= 3"

[tool.poetry.extras]
test = ["pytest", "pytest-cov"]
dev = ["pytest", "pytest-cov"]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = ["error"]
log_cli_level = "INFO"
testpaths = ["tests"]

[tool.coverage]
run.source = ["acbm"]
port.exclude_lines = ['pragma: no cover', '\.\.\.', 'if typing.TYPE_CHECKING:']


[tool.ruff]
src = ["src"]
exclude = []
line-length = 88 # how long you want lines to be

[tool.ruff.format]
docstring-code-format = true # code snippets in docstrings will be formatted

[tool.ruff.lint]
select = [
  "E",
  "F",
  "W",   # flake8
  "B",   # flake8-bugbear
  "I",   # isort
  "ARG", # flake8-unused-arguments
  "C4",  # flake8-comprehensions
  "EM",  # flake8-errmsg
  "ICN", # flake8-import-conventions
  "ISC", # flake8-implicit-str-concat
  "G",   # flake8-logging-format
  "PGH", # pygrep-hooks
  "PIE", # flake8-pie
  "PL",  # pylint
  "RET", # flake8-return
  "RUF", # Ruff-specific
  "SIM", # flake8-simplify
  "UP",  # pyupgrade
  "YTT", # flake8-2020
  "EXE", # flake8-executable
]

ignore = [
  "PLR",    # Design related pylint codes
  "ISC001", # Conflicts with formatter
  "E501",   # Line too long
  "G004",   # Logging statement uses f-string, not necessary here
]
unfixable = [
  "F401", # Would remove unused imports
  "F841", # Would remove unused variables
]
flake8-unused-arguments.ignore-variadic-names = true # allow unused *args/**kwargs
